python: unify PY2 and PY3
